[Feat] Keychain 구현 (추상화)
